+++ src/smash.cc
@@ -538,8 +538,8 @@ int main(int argc, char *argv[]) {
       /* Print only 2->n, n > 1. Do not dump decays, which can be found in
        * decaymodes.txt anyway */
       configuration.merge_yaml("{Collision_Term: {Two_to_One: False}}");
-      initalize_particles_decays_and_tabulations(configuration, version,
-                                                 tabulations_path);
+      initialize_particles_decays_and_tabulations(configuration, version,
+                                                  tabulations_path);
       auto scat_finder = actions_finder_for_dump(configuration);
 
       ignore_simulation_config_values(configuration);
@@ -549,8 +549,8 @@ int main(int argc, char *argv[]) {
       std::exit(EXIT_SUCCESS);
     }
     if (particles_dump_iSS_format) {
-      initalize_particles_decays_and_tabulations(configuration, version,
-                                                 tabulations_path);
+      initialize_particles_decays_and_tabulations(configuration, version,
+                                                  tabulations_path);
       ParticleTypePtrList list;
       list.clear();
       for (const auto &ptype : ParticleType::list_all()) {
@@ -593,8 +593,8 @@ int main(int argc, char *argv[]) {
     }
     if (resonance_dump_activated) {
       // Ignore config values that don't make sense.
-      initalize_particles_decays_and_tabulations(configuration, version,
-                                                 tabulations_path);
+      initialize_particles_decays_and_tabulations(configuration, version,
+                                                  tabulations_path);
       const auto _dummy = ExperimentBase::create(configuration, output_path);
       ignore_simulation_config_values(configuration);
       check_for_unused_config_values(configuration);
@@ -605,8 +605,8 @@ int main(int argc, char *argv[]) {
       std::exit(EXIT_SUCCESS);
     }
     if (cross_section_dump_activated) {
-      initalize_particles_decays_and_tabulations(configuration, version,
-                                                 tabulations_path);
+      initialize_particles_decays_and_tabulations(configuration, version,
+                                                  tabulations_path);
       std::string arg_string(cs_string);
       std::vector<std::string> args = split(arg_string, ',');
       const unsigned int n_arg = args.size();
@@ -690,8 +690,8 @@ int main(int argc, char *argv[]) {
         << "# Date     : " << BUILD_DATE << '\n'
         << configuration.to_string() << '\n';
 
-    initalize_particles_decays_and_tabulations(configuration, version,
-                                               tabulations_path);
+    initialize_particles_decays_and_tabulations(configuration, version,
+                                                tabulations_path);
 
     // Create an experiment
     logg[LMain].trace(SMASH_SOURCE_LOCATION, " create Experiment");